refactor(website): extract LearningCategoryNav component
Split the learning directory sidebar's category nav out of
LearningDirectorySection into its own component, driven by category and
locale props. No visual change.
